# openapi/v1.yaml
openapi: 3.0.3
info:
  title: Procezo API
  version: 1.4.0
paths:
...
  /api/v1/decisions/:
    get:
      operationId: decisions_list
      parameters:
      - in: query
        name: case
        schema:
          type: string
          format: uuid
        required: true
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      tags:
      - decisions
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedDecisionList'
          description: ''
    post:
      operationId: decisions_create
      tags:
      - decisions
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecisionCreate'
        required: true
      security:
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Decision'
          description: ''
...
  /api/v1/demandes/:
    get:
      operationId: demandes_list
      parameters:
      - in: query
        name: case
        schema:
          type: string
          format: uuid
        required: true
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      tags:
      - demandes
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedRequestList'
          description: ''
    post:
      operationId: demandes_create
      tags:
      - demandes
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestCreate'
        required: true
      security:
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Request'
          description: ''
...
  /api/v1/dossiers/:
    get:
      operationId: dossiers_list
      parameters:
      - in: query
        name: ordering
        schema:
          type: string
          enum:
          - -created_at
          - -reference
          - created_at
          - reference
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - in: query
        name: reference
        schema:
          type: string
      - in: query
        name: status
        schema:
          type: string
      - in: query
        name: unit
        schema:
          type: string
          format: uuid
      tags:
      - dossiers
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCaseList'
          description: ''
    post:
      operationId: dossiers_create
      tags:
      - dossiers
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CaseCreate'
        required: true
      security:
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Case'
          description: ''
...
  /api/v1/feuilles/:
    get:
      operationId: feuilles_list
      parameters:
      - in: query
        name: case
        schema:
          type: string
          format: uuid
        required: true
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      tags:
      - feuilles
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedSheetList'
          description: ''
    post:
      operationId: feuilles_create
      tags:
      - feuilles
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SheetInput'
        required: true
      security:
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sheet'
          description: ''
...
  /api/v1/missions/:
    get:
      operationId: missions_list
      parameters:
      - in: query
        name: case
        schema:
          type: string
          format: uuid
        required: true
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      tags:
      - missions
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedMissionList'
          description: ''
    post:
      operationId: missions_create
      tags:
      - missions
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MissionInput'
        required: true
      security:
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mission'
          description: ''
...
  /api/v1/renseignements/:
    get:
      operationId: renseignements_list
      parameters:
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      tags:
      - renseignements
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedIntelligenceList'
          description: ''
    post:
      operationId: renseignements_create
      tags:
      - renseignements
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IntelligenceCreate'
        required: true
      security:
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Intelligence'
          description: ''
...
  /api/v1/statistiques/:
    get:
      operationId: statistics_summary
      parameters:
      - in: query
        name: end
        schema:
          type: string
          format: date
        required: true
      - in: query
        name: start
        schema:
          type: string
          format: date
        required: true
      - in: query
        name: unit
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - statistiques
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Statistics'
          description: ''
...
  /api/v1/transferts-gelec/:
    get:
      operationId: transferts_gelec_list
      parameters:
      - in: query
        name: case
        schema:
          type: string
          format: uuid
        required: true
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      tags:
      - transferts-gelec
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedGelecTransferList'
          description: ''
    post:
      operationId: transferts_gelec_create
      tags:
      - transferts-gelec
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferCreate'
        required: true
      security:
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GelecTransfer'
          description: ''
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GelecTransfer'
          description: ''
...
components:
  schemas:
...
    Case:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        reference:
          type: string
          readOnly: true
        unit:
          type: string
          format: uuid
          readOnly: true
        unit_code:
          type: string
          readOnly: true
        classification:
          type: integer
          readOnly: true
        status:
          allOf:
          - $ref: '#/components/schemas/StatusEnum'
          readOnly: true
        assignee:
          type: integer
          readOnly: true
        assignee_username:
          type: string
          readOnly: true
        next_action:
          type: string
          readOnly: true
        version:
          type: integer
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - assignee
      - assignee_username
      - classification
      - created_at
      - id
      - next_action
      - reference
      - status
      - unit
      - unit_code
      - updated_at
      - version
...
    CaseCreate:
      type: object
      properties:
        unit:
          type: string
          format: uuid
        classification:
          $ref: '#/components/schemas/ClassificationEnum'
        assignee:
          type: integer
          minimum: 1
        next_action:
          type: string
          maxLength: 240
        assignment_reason:
          type: string
          maxLength: 500
      required:
      - assignee
      - assignment_reason
      - classification
      - next_action
      - unit
...
    Decision:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        case:
          type: string
          format: uuid
        author:
          type: integer
        kind:
          $ref: '#/components/schemas/KindEnum'
        reason:
          type: string
          maxLength: 1000
        request_assessment:
          type: string
          format: uuid
          nullable: true
        observation_assessment:
          type: string
          format: uuid
          nullable: true
        replaces:
          type: string
          format: uuid
          nullable: true
        state:
          $ref: '#/components/schemas/DecisionStateEnum'
        version:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        validator:
          type: integer
          nullable: true
        return_comment:
          type: string
          maxLength: 500
        validated_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        prototype_only:
          type: boolean
      required:
      - author
      - case
      - created_at
      - id
      - kind
      - reason
      - updated_at
    DecisionCreate:
      type: object
      properties:
        case:
          type: string
          format: uuid
        case_version:
          type: integer
          minimum: 1
        kind:
          $ref: '#/components/schemas/KindEnum'
        reason:
          type: string
          maxLength: 1000
        request_assessment:
          type: string
          format: uuid
        observation_assessment:
          type: string
          format: uuid
        replaces:
          type: string
          format: uuid
      required:
      - case
      - case_version
      - kind
      - reason
...
    MissionInput:
      type: object
      properties:
        case:
          type: string
          format: uuid
        context:
          type: string
          maxLength: 500
        findings:
          type: string
          maxLength: 4000
        occurred_on:
          type: string
          format: date
        participants:
          type: array
          items:
            type: integer
            minimum: 1
          maxItems: 20
        documents:
          type: array
          items:
            type: string
            format: uuid
          maxItems: 20
      required:
      - case
      - context
      - findings
      - occurred_on
      - participants
...
    PaginatedCaseList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/Case'
...
    PaginatedDecisionList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/Decision'
...
    PaginatedMissionList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/Mission'
...
    PaginatedRequestList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/Request'
...
    PaginatedSheetList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/Sheet'
...
    Request:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        case:
          type: string
          format: uuid
        author:
          type: integer
        target_type:
          $ref: '#/components/schemas/TargetTypeEnum'
        target_name:
          type: string
          maxLength: 240
        represented_name:
          type: string
          maxLength: 240
        subject:
          type: string
          maxLength: 240
        mode:
          $ref: '#/components/schemas/ModeEnum'
        due_on:
          type: string
          format: date
          nullable: true
        state:
          $ref: '#/components/schemas/RequestStateEnum'
        version:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/Item'
          readOnly: true
        acts:
          type: array
          items:
            $ref: '#/components/schemas/Act'
          readOnly: true
        issuance:
          allOf:
          - $ref: '#/components/schemas/Issuance'
          readOnly: true
      required:
      - acts
      - author
      - case
      - created_at
      - id
      - issuance
      - items
      - mode
      - subject
      - target_name
      - target_type
      - updated_at
    RequestCreate:
      type: object
      properties:
        case:
          type: string
          format: uuid
        target_type:
          $ref: '#/components/schemas/TargetTypeEnum'
        target_name:
          type: string
          maxLength: 240
        represented_name:
          type: string
          default: ''
          maxLength: 240
        subject:
          type: string
          maxLength: 240
        mode:
          $ref: '#/components/schemas/ModeEnum'
        due_on:
          type: string
          format: date
          nullable: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/ItemInput'
      required:
      - case
      - items
      - mode
      - subject
      - target_name
      - target_type
...
    Sheet:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        case:
          type: string
          format: uuid
        author:
          type: integer
        mission:
          type: string
          format: uuid
          nullable: true
        origin:
          $ref: '#/components/schemas/OriginEnum'
        recipient_address:
          type: string
          maxLength: 500
        concerned_party:
          type: string
          maxLength: 240
        facts:
          type: string
          maxLength: 4000
        version:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        prototype_only:
          type: boolean
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        observations:
          type: array
          items:
            $ref: '#/components/schemas/Observation'
          readOnly: true
        projects:
          type: array
          items:
            $ref: '#/components/schemas/Project'
          readOnly: true
      required:
      - author
      - case
      - concerned_party
      - created_at
      - facts
      - id
      - observations
      - origin
      - projects
      - recipient_address
      - updated_at
    SheetInput:
      type: object
      properties:
        case:
          type: string
          format: uuid
        mission:
          type: string
          format: uuid
          nullable: true
        origin:
          $ref: '#/components/schemas/OriginEnum'
        recipient_address:
          type: string
          maxLength: 500
        concerned_party:
          type: string
          maxLength: 240
        facts:
          type: string
          maxLength: 4000
        observations:
          type: array
          items:
            $ref: '#/components/schemas/ObservationInput'
      required:
      - case
      - concerned_party
      - facts
      - observations
      - origin
      - recipient_address
...
    TransferCreate:
      type: object
      properties:
        decision:
          type: string
          format: uuid
      required:
      - decision
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid